Using allocatable for dummies arguments is an extension (maybe 200N?) . One of the arguments was passed without the ALLOCATABLE specs and this was giving an error of a non allocated array. Converting the "global" arrays to pointed and keping only the local one as allocatable.
Everytime a unit is requested, once the IO is finished the unit must be given back.

Molden output cannot be embedded in an if like (IF (ionode)..) this does not guarantee
that on a specific machine the output will really be there. Once you've asked for the IO
unit you know who will print the info just checking if the unit is > 0 on a local processor.

TEST_TYPE 8 does not work because the output is different. Changed the output and added also
another TEST_TYPE for this specific run.

Strange behavior observed in parallel.